Refactor: Per-block recovery + opt-in tracing across renderer and component
Two related shifts that align dev-time perf with production and make
runtime errors loud by default.

Per-block recovery:
defineBlock no longer wraps every hook in try/catch. Recovery is opt-in:
blocks that define an `error` hook get the wrapping; blocks without let
throws propagate. Aligns with "page doesn't load is the strongest signal
that something is broken" — silent recovery + visible-but-stale UI is
the worst failure mode for agents and humans iterating on changes.
async is the canonical opt-in (its `{error as e}` branch needs the
recovery hook). Other blocks let throws bubble.

Destroy stays defensively wrapped — a throw there strands sibling
cleanup and matters more for correctness than for visibility. Logs
through console.error (always-on) rather than the structured emitter.

Tracing now opt-in:
Replaces isDevelopment guards on the structured error log with an
isTracing() check from packages/renderer/src/helpers.js. Default off
even in dev. Together with the reactivity-side setTracing decoupling
(prior commit), dev-time rendering matches production performance for
debug features nobody uses on every tick.

  import { setTracing } from '@semantic-ui/component';   // both layers
  import { setTracing } from '@semantic-ui/reactivity';  // reactions only
  import { setTracing } from '@semantic-ui/renderer';    // blocks only

Naming convention: same function name in each package, scoped by import
path. Consumer ergonomics: one obvious import from /component flips
everything; finer-grained imports for layered debugging.

Renamed reactivity/src/dev-tracing.js to helpers.js to match the
@semantic-ui/query/src/helpers.js convention for small package-level
utilities. Renderer follows the same pattern.

1097 tests pass (86 reactivity + 932 renderer + 79 component). The
single failing reactivity debug test and the two failing renderer
emitter tests now wrap with setTracing(true)/(false) to opt in to the
tracing they assert on.
