+++ packages/component/src/index.js
@@ -7,4 +7,5 @@ export { getComponent, hasComponent, registerComponent } from './component-regis
 export { defineComponent } from './define-component.js';
 export { WebComponentBase } from './engines/native/base.js';
 export { NativeEngine } from './engines/native/register.js';
+export { setTracing } from './helpers.js';
 export { renderToString } from './render-to-string.js';
