+++ packages/reactivity/src/index.js
@@ -1,5 +1,5 @@
 export { Dependency } from './dependency.js';
-export { isTracing, setTracing } from './dev-tracing.js';
+export { isTracing, setTracing } from './helpers.js';
 export { Reaction } from './reaction.js';
 export { Scheduler } from './scheduler.js';
 export { Signal } from './signal.js';
