+++ packages/renderer/src/helpers.js
@@ -1,0 +1,7 @@
+let tracing = false;
+
+export const setTracing = (enabled) => {
+  tracing = !!enabled;
+};
+
+export const isTracing = () => tracing;
